+++ src/cmd/hack.go
@@ -1,8 +1,6 @@
 package cmd
 
 import (
-	"errors"
-
 	"github.com/Originate/git-town/src/git"
 	"github.com/Originate/git-town/src/script"
 	"github.com/Originate/git-town/src/steps"
@@ -27,8 +25,8 @@ and brings over all uncommitted changes to the new feature branch.
 
 Additionally, when there is a remote upstream,
 the main branch is synced with its upstream counterpart.
-This can be disabled by toggling the "hack-push-flag" configuration:
-$ git town hack-push-flag false`,
+This can be disabled by toggling the "new-branch-push-flag" configuration:
+$ git town new-branch-push-flag false`,
 	Run: func(cmd *cobra.Command, args []string) {
 		steps.Run(steps.RunOptions{
 			CanSkip:              func() bool { return false },
@@ -44,12 +42,14 @@ $ git town hack-push-flag false`,
 			},
 		})
 	},
-	PreRunE: func(cmd *cobra.Command, args []string) error {
-		if len(args) == 0 && !abortFlag && !continueFlag {
-			return errors.New("no branch name provided")
+	Args: func(cmd *cobra.Command, args []string) error {
+		if abortFlag || continueFlag {
+			return cobra.NoArgs(cmd, args)
 		}
+		return cobra.ExactArgs(1)(cmd, args)
+	},
+	PreRunE: func(cmd *cobra.Command, args []string) error {
 		return util.FirstError(
-			validateMaxArgsFunc(args, 1),
 			git.ValidateIsRepository,
 			validateIsConfigured,
 		)
@@ -67,9 +67,9 @@ func getHackConfig(args []string) (result hackConfig) {
 
 func getHackStepList(config hackConfig) (result steps.StepList) {
 	mainBranchName := git.GetMainBranch()
-	result.AppendList(steps.GetSyncBranchSteps(mainBranchName))
+	result.AppendList(steps.GetSyncBranchSteps(mainBranchName, true))
 	result.Append(&steps.CreateAndCheckoutBranchStep{BranchName: config.TargetBranch, ParentBranchName: mainBranchName})
-	if git.HasRemote("origin") && git.ShouldHackPush() && !git.IsOffline() {
+	if git.HasRemote("origin") && git.ShouldNewBranchPush() && !git.IsOffline() {
 		result.Append(&steps.CreateTrackingBranchStep{BranchName: config.TargetBranch})
 	}
 	result.Wrap(steps.WrapOptions{RunInGitRoot: true, StashOpenChanges: true})