+++ src/cmd/kill.go
@@ -41,9 +41,14 @@ Does not delete perennial branches nor the main branch.`,
 			},
 		})
 	},
+	Args: func(cmd *cobra.Command, args []string) error {
+		if undoFlag {
+			return cobra.NoArgs(cmd, args)
+		}
+		return cobra.MaximumNArgs(1)(cmd, args)
+	},
 	PreRunE: func(cmd *cobra.Command, args []string) error {
 		return util.FirstError(
-			validateMaxArgsFunc(args, 1),
 			git.ValidateIsRepository,
 			validateIsConfigured,
 		)
@@ -94,7 +99,6 @@ func getKillStepList(config killConfig) (result steps.StepList) {
 			result.Append(&steps.SetParentBranchStep{BranchName: child, ParentBranchName: targetBranchParent})
 		}
 		result.Append(&steps.DeleteParentBranchStep{BranchName: config.TargetBranch})
-		result.Append(&steps.DeleteAncestorBranchesStep{})
 	} else if !git.IsOffline() {
 		result.Append(&steps.DeleteRemoteBranchStep{BranchName: config.TargetBranch, IsTracking: false})
 	} else {