+++ src/cmd/rename_branch.go
@@ -1,7 +1,6 @@
 package cmd
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/Originate/git-town/src/git"
@@ -56,12 +55,14 @@ When run on a perennial branch
 			},
 		})
 	},
-	PreRunE: func(cmd *cobra.Command, args []string) error {
-		if len(args) == 0 && !undoFlag {
-			return errors.New("Too few arguments")
+	Args: func(cmd *cobra.Command, args []string) error {
+		if undoFlag {
+			return cobra.NoArgs(cmd, args)
 		}
+		return cobra.RangeArgs(1, 2)(cmd, args)
+	},
+	PreRunE: func(cmd *cobra.Command, args []string) error {
 		return util.FirstError(
-			validateMaxArgsFunc(args, 2),
 			git.ValidateIsRepository,
 			validateIsConfigured,
 		)
@@ -107,7 +108,6 @@ func getRenameBranchStepList(config renameBranchConfig) (result steps.StepList) 
 	for _, child := range git.GetChildBranches(config.OldBranchName) {
 		result.Append(&steps.SetParentBranchStep{BranchName: child, ParentBranchName: config.NewBranchName})
 	}
-	result.Append(&steps.DeleteAncestorBranchesStep{})
 	if git.HasTrackingBranch(config.OldBranchName) && !git.IsOffline() {
 		result.Append(&steps.CreateTrackingBranchStep{BranchName: config.NewBranchName})
 		result.Append(&steps.DeleteRemoteBranchStep{BranchName: config.OldBranchName, IsTracking: true})