+++ src/steps/load_state.go
@@ -83,8 +83,6 @@ func deserializeStep(serializedStep SerializedStep) Step {
 		step := CreateTrackingBranchStep{}
 		exit.If(json.Unmarshal(serializedStep.Data, &step))
 		return &step
-	case "*DeleteAncestorBranchesStep":
-		return &DeleteAncestorBranchesStep{}
 	case "*DeleteLocalBranchStep":
 		step := DeleteLocalBranchStep{}
 		exit.If(json.Unmarshal(serializedStep.Data, &step))
@@ -109,8 +107,6 @@ func deserializeStep(serializedStep SerializedStep) Step {
 		step := MergeBranchStep{}
 		exit.If(json.Unmarshal(serializedStep.Data, &step))
 		return &step
-	case "*MergeTrackingBranchStep":
-		return &MergeTrackingBranchStep{}
 	case "*NoOpStep":
 		return &NoOpStep{}
 	case "*PreserveCheckoutHistoryStep":
@@ -133,8 +129,6 @@ func deserializeStep(serializedStep SerializedStep) Step {
 		step := RebaseBranchStep{}
 		exit.If(json.Unmarshal(serializedStep.Data, &step))
 		return &step
-	case "*RebaseTrackingBranchStep":
-		return &RebaseTrackingBranchStep{}
 	case "*RemoveFromPerennialBranches":
 		step := RemoveFromPerennialBranches{}
 		exit.If(json.Unmarshal(serializedStep.Data, &step))